refactor(workflows): address review comments from PR #1252
Three small follow-ups to the execution_service.ts refactor, all
identified in the review of PR #1252:

1. **Driver fallback preserves the definition tier.** The
   `ctx.driverPlan?.withDefinition(...) ?? { driver: "raw" }` shape
   silently dropped the definition's driver when no plan was passed —
   a subtle behaviour change vs the original `resolveDriverConfig(...,
   { driver, driverConfig }, ...)` which always honoured the definition
   tier. Production isn't affected (runStep always builds a DriverPlan)
   but tests / direct invokers of DefaultStepExecutor are. Routes
   through an empty DriverPlan when none is provided so the definition
   tier is still applied:

     const resolved = (ctx.driverPlan ?? new DriverPlan({})).withDefinition({...})

2. **Single shared CEL evaluator interface.** `CelExpressionEvaluator`
   (in expression_evaluators.ts, async-only) and `CelEvaluatorRuntime`
   (in for_each_expansion_service.ts, sync+async) were the same
   abstraction with two surfaces and two definitions. Consolidate to a
   single `CelExpressionEvaluator` in `src/domain/expressions/cel_runtime.ts`
   with both methods. Consumers that only need one method are still
   served — TypeScript's structural typing handles the broader shape
   for narrower uses.

3. **JSDoc reordering in expression_evaluators.ts.** The class-level
   doc for `WorkflowExpressionEvaluator` sat above the
   `WorkflowEvaluationResult` interface that was added between the doc
   block and the class. Move the interface above the doc block so the
   class description sits directly above its target.

Behaviour-preserving:
- Driver fallback now matches pre-DriverPlan behaviour exactly
- CEL interface consolidation is purely structural — same shape,
  one location
- JSDoc move is presentational only
- Verified: deno check / lint / fmt / test (5045 passed),
  /tmp/swamp-verification check OK against baseline
